7-to_kv.py

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -8,6 +8,6 @@ annotated as a float.
 from typing import Union, Tuple
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """ ... """
-    return (k, float(v**2))
+    return (k, float(v ** 2))
